Adding after local dev work
This is probably the 3rd or 4th attempt at starting this project over, but I like where it is at now. I'm ready to commit this so that I can continue enhancing it--possibly with a different API. I'm excluding the node_modules package, but I'm including files labeled with "practice" so I can keep track of my notes and past work on the versions before the one I'm submitting.